Update build.yml schedule to run on the 3rd of every month
When specifying a day of the month and a day of the week to the chron syntax, it will treat it as an 'or' expression instead of an 'and' expression, as I intended. I was searching for all days of the month that could be in the 3rd week and on Wednesday, but the syntax would've treated this as every Wednesday or every day of the 3rd week of the month. It would be complicated to fix this with conditionals or other methods while keeping the ability to activate on pushes, so I've decided instead to update the schedule to only run on the 3rd of the month.